+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - #852 Implement patchedast handlers for TypeAlias 
 - #853 Implement patchedast handlers TypeVar
 - #847 Avoid printing autoimport syntax errors (@yangfan-yf-yf)
+- #819 supports MatchOr, MatchSequence, MatchStar (@jheld)
 
 # Release 1.14.0